v0.1.0 polish: fix CI, harden runner, friendlier API, README rewrite
- CI fix: re-include gui/assets/*.png and docs/assets/*.png via .gitignore
  negation and add the four PNG files. Was blocking the wheel build on the
  CI runner because force-include couldn't find files that were never in
  the repo.
- simulation.py: tolerate hyphenated BC types ("velocity-inlet") and
  parameter_path with a redundant trailing ".value"; auto-resolve bc_type
  when set_inputs got the bare (lo, hi) tuple form.
- Project.connect_fluent now accepts precision/processor_count/dimension/
  use_gui kwargs (mixing_elbow is single-precision; needed an override).
- Project.run_simulations gains verbose=True for a built-in per-sim print
  callback. Two elbow examples + integration test use it.
- README rewritten for the dual GUI/API surface, docs link, video slot.

+++ docs/examples/discovering_bcs.py
@@ -19,7 +19,7 @@ CASE_FILE = examples.download_file("mixing_elbow.cas.h5", "pyfluent/mixing_elbow
 
 print(f"Loading {CASE_FILE}...\n")
 solver = pyfluent.launch_fluent(
-    precision="double", processor_count=4, dimension=3, mode="solver",
+    precision="single", processor_count=4, dimension=3, mode="solver",
 )
 solver.settings.file.read_case(file_name=CASE_FILE)
